refactor: Move Helpers to `lib` directory
patch: Remove redundant importing of auto-injected vue helpers
