+++ src/components/app/header.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import dayjs from 'dayjs';
 
-import { controlLinks, navLinks } from '@/constants';
+import { controlLinks, navLinks } from '@/lib/constants';
 </script>
 
 <template>
@@ -18,7 +18,7 @@ import { controlLinks, navLinks } from '@/constants';
 		<div id="app-header_control-panel">
 			<ul>
 				<li v-for="(opt) in controlLinks" :key="opt.label">
-					<img :src="`/images/${opt.img}`" :alt="opt.label">
+					<img :src="`/images/${opt.imgSrc}`" :alt="opt.label">
 				</li>
 			</ul>
 			<time>{{ dayjs().format('ddd MMM D h:mm A') }}</time>
