+++ src/components/app/text-unpacker.vue
@@ -1,6 +1,4 @@
 <script lang='ts' setup>
-import { defineProps, withDefaults } from 'vue';
-
 withDefaults(defineProps<{
 	text: string;
 	className?: string;
