--- src/constants/index.ts
@@ -1,21 +1,0 @@
-export const navLinks = [
-	{ label: 'Portfolio', link: '/' },
-	{ label: 'Contact', link: '/' },
-	{ label: 'Projects', link: '/' },
-];
-
-export const controlLinks = [
-	{ label: 'wifi', img: 'icons/wifi.svg' },
-	{ label: 'search', img: 'icons/search.svg' },
-	{ label: 'user', img: 'icons/user.svg' },
-	{ label: 'mode', img: 'icons/mode.svg' },
-];
-
-export const FONT_WEIGHTS = {
-	subtitle: { min: 100, max: 900, default: 100 },
-	title: { min: 400, max: 900, default: 400 },
-};
-
-export const DOCK_APPS = [
-	{ id: 'finder', label: 'Finder', icon: 'app-icons/finder.svg', canOpen: true },
-];
