+++ src/lib/constants/index.ts
@@ -1,0 +1,28 @@
+import type { DockApp } from '@/lib/types';
+
+export const navLinks = [
+	{ label: 'Portfolio', link: '/' },
+	{ label: 'Contact', link: '/' },
+	{ label: 'Projects', link: '/' },
+];
+
+export const controlLinks = [
+	{ label: 'wifi', imgSrc: 'icons/wifi.svg' },
+	{ label: 'search', imgSrc: 'icons/search.svg' },
+	{ label: 'user', imgSrc: 'icons/user.svg' },
+	{ label: 'mode', imgSrc: 'icons/mode.svg' },
+];
+
+export const FONT_WEIGHTS = {
+	subtitle: { min: 100, max: 900, default: 100 },
+	title: { min: 400, max: 900, default: 400 },
+};
+
+export const DOCK_APPS: DockApp[] = [
+	{ id: 'finder', label: 'Portfolio', imgSrc: 'app-icons/finder.png', canOpen: true },
+	{ id: 'safari', label: 'Articles', imgSrc: 'app-icons/safari.png', canOpen: true },
+	{ id: 'photos', label: 'Gallery', imgSrc: 'app-icons/photos.png', canOpen: true },
+	{ id: 'contact', label: 'Contact', imgSrc: 'app-icons/contact.png', canOpen: true },
+	{ id: 'terminal', label: 'Skills', imgSrc: 'app-icons/terminal.png', canOpen: true },
+	{ id: 'trash', label: 'Archive', imgSrc: 'app-icons/trash.png', canOpen: false },
+];
