+++ src/lib/types/index.ts
@@ -1,0 +1,1 @@
+export type DockApp = { id: string; label: string; imgSrc: string; canOpen: boolean };
